+++ 01 Hardware/01 Lidar/03 Software/01 source/LidarTest/LidarTest/timerUnit.h
@@ -38,7 +38,7 @@ typedef struct{
 	uint8_t state:1;
 }timerStatus_t;
 
-extern uint8_t timerInit(uint8_t rezConfig, uint8_t resolution);
+extern uint8_t timerInit(uint8_t rezConfig, uint16_t resolution);
 extern void timer_setState(uint8_t state);
 extern void timer_setCounter(uint32_t value);
 extern const int16_t timer_getCounter();
